refactor(actor-runtime): split spec_actor into a directory module
spec_actor.rs reached 899 lines after the ARN-179 fix; the repo limit is
500. Content-preserving split: mod.rs (actor, messages, state, routing),
effects.rs (effect vocabulary: validation + application), tests.rs. The
public path spec_actor::validate_effect_support is re-exported unchanged.
