--- crates/temper-actor-runtime/src/spec_actor.rs
@@ -1,899 +1,0 @@
-//! SpecDrivenActor — implements the Actor trait backed by an IOA spec.
-//!
-//! Specs describe state machines (states, transitions, guards, effects).
-//! The routing is external — reaction rules wire emit effects to target actors.
-//!
-//! # Architecture
-//!
-//! - Spec → TransitionTable (via temper-jit)
-//! - Reaction rules → routing map (emit name → target actor type)
-//! - handle(): evaluate table → apply effects → route emits via ctx.tell()
-//!
-//! # Message protocol
-//!
-//! Actors communicate via `SpecMessage { action, params }`:
-//! - `action`: the action/emit name (e.g., "PrepareContext")
-//! - `params`: JSON-encoded params (empty for actions with no params)
-
-use std::collections::{BTreeMap, HashMap};
-
-use temper_jit::table::{EvalContext, TransitionTable};
-use temper_runtime::reaction::ReactionRule;
-use temper_spec::automaton::Automaton;
-
-use crate::actor::{Actor, ActorContext, ActorError, ActorHandle, Message};
-
-// ─── SpecMessage ─────────────────────────────────────────────────────────────
-
-/// Generic message for spec-driven actor communication.
-/// The action name matches the IOA spec action/emit name.
-#[derive(Clone, PartialEq, prost::Message)]
-pub struct SpecMessage {
-    /// The action/emit name (e.g., "PrepareContext", "ToolCallBatchRequested").
-    #[prost(string, tag = "1")]
-    pub action: String,
-    /// JSON-encoded params (empty bytes for parameterless actions).
-    #[prost(bytes, tag = "2")]
-    pub params: Vec<u8>,
-}
-
-impl SpecMessage {
-    pub fn new(action: impl Into<String>) -> Self {
-        Self {
-            action: action.into(),
-            params: Vec::new(),
-        }
-    }
-
-    pub fn with_params(action: impl Into<String>, params: serde_json::Value) -> Self {
-        Self {
-            action: action.into(),
-            params: serde_json::to_vec(&params).unwrap_or_default(),
-        }
-    }
-}
-
-// ─── Actor state ─────────────────────────────────────────────────────────────
-
-/// Serializable state for spec-driven actors.
-#[derive(Debug, Clone, serde::Serialize, serde::Deserialize, Default)]
-pub struct SpecActorState {
-    pub status: String,
-    #[serde(default)]
-    pub counters: BTreeMap<String, usize>,
-    #[serde(default)]
-    pub booleans: BTreeMap<String, bool>,
-    #[serde(default)]
-    pub lists: BTreeMap<String, Vec<String>>,
-    /// Arbitrary extra data — used to thread params through the reaction chain.
-    /// SpecDrivenActor stores the last incoming message params here so integrations
-    /// can read them from the trigger message.
-    #[serde(default)]
-    pub fields: serde_json::Value,
-}
-
-impl SpecActorState {
-    fn to_eval_context(&self) -> EvalContext {
-        let mut ctx = EvalContext::default();
-        for (k, v) in &self.counters {
-            ctx.counters.insert(k.clone(), *v);
-        }
-        for (k, v) in &self.booleans {
-            ctx.booleans.insert(k.clone(), *v);
-        }
-        for (k, v) in &self.lists {
-            ctx.lists.insert(k.clone(), v.clone());
-        }
-        ctx
-    }
-}
-
-// ─── Routing map builder ─────────────────────────────────────────────────────
-
-/// Build per-actor routing maps from reaction rules.
-///
-/// Returns `HashMap<actor_type, HashMap<emit_name, (target_actor_type, target_action)>>`.
-pub fn build_routing_maps(
-    rules: &[ReactionRule],
-) -> HashMap<String, HashMap<String, (String, String)>> {
-    let mut maps: HashMap<String, HashMap<String, (String, String)>> = HashMap::new();
-
-    for rule in rules {
-        if let Some(emit_name) = &rule.when.action {
-            maps.entry(rule.when.entity_type.clone())
-                .or_default()
-                .insert(
-                    emit_name.clone(),
-                    (rule.then.entity_type.clone(), rule.then.action.clone()),
-                );
-        }
-    }
-
-    maps
-}
-
-/// Build a single actor's routing map from a reaction registry.
-pub fn build_actor_routing(
-    actor_type: &str,
-    rules: &[ReactionRule],
-) -> HashMap<String, (String, String)> {
-    rules
-        .iter()
-        .filter(|r| r.when.entity_type == actor_type)
-        .filter_map(|r| {
-            r.when.action.as_ref().map(|emit| {
-                (
-                    emit.clone(),
-                    (r.then.entity_type.clone(), r.then.action.clone()),
-                )
-            })
-        })
-        .collect()
-}
-
-// ─── Effect vocabulary support ───────────────────────────────────────────────
-
-/// Validate that every effect in the compiled table is executable by this
-/// runtime. This is the single source of truth for the Postgres actor
-/// runtime's effect vocabulary (ARN-179): specs are rejected here, at
-/// construction, instead of having unsupported effects silently dropped
-/// during effect application.
-///
-/// The match is exhaustive on purpose — a new [`temper_jit::table::Effect`]
-/// variant fails compilation here, forcing an explicit support decision.
-pub fn validate_effect_support(
-    table: &TransitionTable,
-    routing: &HashMap<String, (String, String)>,
-) -> Result<(), String> {
-    use temper_jit::table::Effect;
-
-    for rule in &table.rules {
-        for effect in &rule.effects {
-            match effect {
-                Effect::SetState(_)
-                | Effect::IncrementItems
-                | Effect::DecrementItems
-                | Effect::IncrementCounter(_)
-                | Effect::IncrementCounterByParam { .. }
-                | Effect::DecrementCounter(_)
-                | Effect::DecrementCounterByParam { .. }
-                | Effect::SetCounterFromParam { .. }
-                | Effect::SetBool { .. }
-                | Effect::ListAppend(_)
-                | Effect::ListRemoveAt(_)
-                // Emits are synthesized for every action; one without a
-                // reaction rule simply has no listener.
-                | Effect::EmitEvent(_) => {}
-                Effect::Custom(trigger) => {
-                    if !routing.contains_key(trigger.as_str()) {
-                        return Err(format!(
-                            "action {:?} uses trigger effect {trigger:?} with no reaction \
-                             routing; wire a reaction rule for it or remove the trigger",
-                            rule.name
-                        ));
-                    }
-                }
-                Effect::ScheduleAction { .. } => {
-                    return Err(unsupported_effect(&rule.name, "schedule"));
-                }
-                Effect::ScheduleAtAction { .. } => {
-                    return Err(unsupported_effect(&rule.name, "schedule_at"));
-                }
-                Effect::SpawnEntity { .. } => {
-                    return Err(unsupported_effect(&rule.name, "spawn"));
-                }
-            }
-        }
-    }
-    Ok(())
-}
-
-fn unsupported_effect(action: &str, effect_type: &str) -> String {
-    format!(
-        "action {action:?} uses effect type {effect_type:?}, which the Postgres actor \
-         runtime cannot execute (it has no delayed delivery or per-entity spawning)"
-    )
-}
-
-// ─── SpecDrivenActor ─────────────────────────────────────────────────────────
-
-/// An Actor implementation driven by an IOA spec + reaction routing.
-///
-/// - State machine transitions validated by the TransitionTable
-/// - Emit effects routed to sibling actors via ctx.tell()
-/// - Trigger effects sent to integration actors via ctx.tell()
-pub struct SpecDrivenActor {
-    /// Actor type name (e.g., "Agent", "ContextManager").
-    name: String,
-    /// TransitionTable compiled from the IOA spec.
-    table: TransitionTable,
-    /// Initial state (from spec's initial state + variable declarations).
-    init_state: SpecActorState,
-    /// Routing map: emit/trigger name → (target actor type, target action).
-    routing: HashMap<String, (String, String)>,
-    /// Leaked static refs for subscriptions() return.
-    subscriptions_static: Vec<&'static str>,
-}
-
-impl SpecDrivenActor {
-    /// Create from an IOA TOML source + routing map.
-    ///
-    /// Returns an error if the spec fails to parse or uses effects this
-    /// runtime cannot execute (see [`validate_effect_support`]).
-    pub fn from_ioa(
-        ioa_source: &str,
-        routing: HashMap<String, (String, String)>,
-    ) -> Result<Self, String> {
-        let automaton = temper_spec::parse_automaton(ioa_source)
-            .map_err(|e| format!("failed to parse spec: {e}"))?;
-        Self::from_automaton(&automaton, ioa_source, routing)
-    }
-
-    /// Create from a pre-parsed Automaton + routing map.
-    ///
-    /// Returns an error if the compiled transition table uses effects this
-    /// runtime cannot execute (see [`validate_effect_support`]). Rejecting at
-    /// construction replaces the former silent catch-all in effect
-    /// application, which dropped unsupported effects at runtime (ARN-179).
-    pub fn from_automaton(
-        automaton: &Automaton,
-        ioa_source: &str,
-        routing: HashMap<String, (String, String)>,
-    ) -> Result<Self, String> {
-        let name = automaton.automaton.name.clone();
-        let table = TransitionTable::from_ioa_source(ioa_source);
-        validate_effect_support(&table, &routing)?;
-
-        // Build initial state from spec variables.
-        let mut init_state = SpecActorState {
-            status: automaton.automaton.initial.clone(),
-            ..Default::default()
-        };
-        for var in &automaton.state {
-            match var.var_type.as_str() {
-                "counter" => {
-                    let v: usize = var.initial.parse().unwrap_or(0);
-                    init_state.counters.insert(var.name.clone(), v);
-                }
-                "bool" => {
-                    let v: bool = var.initial.parse().unwrap_or(false);
-                    init_state.booleans.insert(var.name.clone(), v);
-                }
-                "list" | "set" => {
-                    init_state.lists.insert(var.name.clone(), Vec::new());
-                }
-                _ => {}
-            }
-        }
-
-        // Input actions are the message types this actor accepts.
-        // NOTE: Box::leak is intentional — actors are singletons, never dropped.
-        let subscriptions_static: Vec<&'static str> = automaton
-            .actions
-            .iter()
-            .filter(|a| a.kind == "input")
-            .map(|a| &*Box::leak(a.name.clone().into_boxed_str()))
-            .collect();
-
-        Ok(Self {
-            name,
-            table,
-            init_state,
-            routing,
-            subscriptions_static,
-        })
-    }
-
-    /// Which message types this actor accepts.
-    pub fn subscription_strings(&self) -> &[&'static str] {
-        &self.subscriptions_static
-    }
-
-    /// The routing map (emit name → target actor type).
-    pub fn routing(&self) -> &HashMap<String, (String, String)> {
-        &self.routing
-    }
-}
-
-#[async_trait::async_trait]
-impl Actor for SpecDrivenActor {
-    fn actor_type(&self) -> &str {
-        &self.name
-    }
-
-    fn initial_state(&self) -> Vec<u8> {
-        serde_json::to_vec(&self.init_state).unwrap_or_default()
-    }
-
-    async fn handle(
-        &self,
-        ctx: &ActorContext,
-        state: &mut Vec<u8>,
-        message: &Message,
-    ) -> Result<(), ActorError> {
-        // 1. Deserialize state.
-        let mut actor_state: SpecActorState = if state.is_empty() {
-            self.init_state.clone()
-        } else {
-            serde_json::from_slice(state)
-                .map_err(|e| ActorError::HandlerFailed(format!("state deser: {e}")))?
-        };
-
-        // 2. Resolve action name + params.
-        // If the message carries a SpecMessage, extract the action from its payload.
-        // This handles both direct SpecMessage sends and raw action-name messages.
-        let spec_msg = if message.message_type.ends_with("SpecMessage") {
-            message.decode::<SpecMessage>().ok()
-        } else {
-            None
-        };
-        let action = spec_msg
-            .as_ref()
-            .filter(|m| !m.action.is_empty())
-            .map(|m| m.action.clone())
-            .unwrap_or_else(|| message.message_type.clone());
-
-        // Store incoming params in state.fields so integrations can read them.
-        // Merge non-empty params into fields to preserve context from prior steps
-        // (e.g. child Process keeps parent_pid while later messages add user_prompt/response).
-        // For a new user turn, clear transient scratchpad fields from prior turns.
-        if self.name == "Process"
-            && matches!(action.as_str(), "StartProcess" | "SendInput")
-            && let Some(obj) = actor_state.fields.as_object_mut()
-        {
-            for key in [
-                "tool_calls",
-                "tool_results",
-                "child_result",
-                "response",
-                "error",
-            ] {
-                obj.remove(key);
-            }
-        }
-
-        let params: serde_json::Value = spec_msg
-            .as_ref()
-            .filter(|m| !m.params.is_empty())
-            .and_then(|m| serde_json::from_slice(&m.params).ok())
-            .unwrap_or(serde_json::Value::Null);
-
-        if !params.is_null() && !params.as_object().is_some_and(|o| o.is_empty()) {
-            match (actor_state.fields.as_object_mut(), params.as_object()) {
-                (Some(existing), Some(new_fields)) => {
-                    for (k, v) in new_fields {
-                        existing.insert(k.clone(), v.clone());
-                    }
-                }
-                _ => actor_state.fields = params.clone(),
-            }
-        }
-
-        let eval_ctx = actor_state.to_eval_context();
-
-        // 2. Evaluate transition table.
-        let result = self
-            .table
-            .evaluate_ctx(&actor_state.status, &eval_ctx, &action);
-
-        match result {
-            Some(r) if r.success => {
-                let from_status = actor_state.status.clone();
-
-                // 3. Apply effects — may include SetState.
-                for effect in &r.effects {
-                    self.apply_effect(&mut actor_state, effect, &params, ctx)
-                        .await?;
-                }
-
-                // 4. Apply state transition fallback (if no SetState effect fired).
-                if actor_state.status == from_status && !r.new_state.is_empty() {
-                    actor_state.status = r.new_state.clone();
-                }
-
-                tracing::info!(
-                    actor = %self.name,
-                    action = %action,
-                    new_state = %actor_state.status,
-                    "transition"
-                );
-            }
-            Some(_) => {
-                tracing::warn!(
-                    actor = %self.name,
-                    action = %action,
-                    status = %actor_state.status,
-                    "action not valid from current state"
-                );
-            }
-            None => {
-                tracing::warn!(
-                    actor = %self.name,
-                    action = %action,
-                    "unknown action"
-                );
-            }
-        }
-
-        // 5. Serialize state back.
-        *state = serde_json::to_vec(&actor_state)
-            .map_err(|e| ActorError::HandlerFailed(format!("state ser: {e}")))?;
-
-        Ok(())
-    }
-}
-
-impl SpecDrivenActor {
-    /// Apply one transition effect to actor state.
-    ///
-    /// The match is exhaustive — no catch-all — so every
-    /// [`temper_jit::table::Effect`] variant has an explicit outcome and a
-    /// new variant fails compilation instead of being silently dropped
-    /// (ARN-179). Param-driven semantics mirror the canonical executor in
-    /// `temper-server::entity_actor::effects::apply_effects`.
-    async fn apply_effect(
-        &self,
-        state: &mut SpecActorState,
-        effect: &temper_jit::table::Effect,
-        params: &serde_json::Value,
-        ctx: &ActorContext,
-    ) -> Result<(), ActorError> {
-        match effect {
-            temper_jit::table::Effect::SetState(s) => {
-                state.status = s.clone();
-            }
-            temper_jit::table::Effect::IncrementItems => {
-                *state.counters.entry("items".into()).or_default() += 1;
-            }
-            temper_jit::table::Effect::IncrementCounter(var) => {
-                *state.counters.entry(var.clone()).or_default() += 1;
-            }
-            temper_jit::table::Effect::IncrementCounterByParam { var, param } => {
-                let delta = counter_delta_from_params(params, param);
-                *state.counters.entry(var.clone()).or_default() += delta;
-            }
-            temper_jit::table::Effect::DecrementItems => {
-                let c = state.counters.entry("items".into()).or_default();
-                *c = c.saturating_sub(1);
-            }
-            temper_jit::table::Effect::DecrementCounter(var) => {
-                let c = state.counters.entry(var.clone()).or_default();
-                *c = c.saturating_sub(1);
-            }
-            temper_jit::table::Effect::DecrementCounterByParam { var, param } => {
-                let delta = counter_delta_from_params(params, param);
-                let c = state.counters.entry(var.clone()).or_default();
-                *c = c.saturating_sub(delta);
-            }
-            temper_jit::table::Effect::SetCounterFromParam { var, param } => {
-                let parsed = params
-                    .get(param)
-                    .and_then(|v| {
-                        v.as_u64()
-                            .or_else(|| v.as_i64().and_then(|n| u64::try_from(n).ok()))
-                    })
-                    .and_then(|n| usize::try_from(n).ok());
-                match parsed {
-                    Some(value) => {
-                        state.counters.insert(var.clone(), value);
-                    }
-                    None => tracing::warn!(
-                        actor = %self.name,
-                        counter = %var,
-                        param = %param,
-                        "set_counter_from_param skipped because param was missing or not a non-negative integer"
-                    ),
-                }
-            }
-            temper_jit::table::Effect::SetBool { var, value } => {
-                state.booleans.insert(var.clone(), *value);
-            }
-            temper_jit::table::Effect::ListAppend(var) => {
-                if let Some(val) = params.get(var).and_then(|v| v.as_str()) {
-                    state
-                        .lists
-                        .entry(var.clone())
-                        .or_default()
-                        .push(val.to_string());
-                }
-            }
-            temper_jit::table::Effect::ListRemoveAt(var) => {
-                let index_key = format!("{var}_index");
-                if let Some(idx) = params.get(&index_key).and_then(|v| v.as_u64()) {
-                    let list = state.lists.entry(var.clone()).or_default();
-                    let idx = idx as usize;
-                    if idx < list.len() {
-                        list.remove(idx);
-                    }
-                }
-            }
-            temper_jit::table::Effect::EmitEvent(emit_name) => {
-                if let Some((target_type, target_action)) = self.routing.get(emit_name.as_str()) {
-                    tracing::info!(actor=%self.name, emit=%emit_name, target=%target_type, target_action=%target_action, "routing emit");
-                    let target =
-                        ActorHandle::new(ctx.self_handle().namespace.clone(), target_type.clone());
-                    ctx.tell(
-                        &target,
-                        SpecMessage::with_params(target_action.clone(), state.fields.clone()),
-                    )
-                    .await;
-                } else {
-                    tracing::warn!(
-                        actor = %self.name,
-                        emit = %emit_name,
-                        "no routing for emit (no reaction rule)"
-                    );
-                }
-            }
-            temper_jit::table::Effect::Custom(trigger_name) => {
-                let Some((target_type, target_action)) = self.routing.get(trigger_name.as_str())
-                else {
-                    // Unreachable: construction rejects unrouted triggers.
-                    return Err(ActorError::HandlerFailed(format!(
-                        "trigger {trigger_name:?} has no reaction routing"
-                    )));
-                };
-                tracing::info!(actor=%self.name, trigger=%trigger_name, target=%target_type, target_action=%target_action, "routing trigger");
-                let target =
-                    ActorHandle::new(ctx.self_handle().namespace.clone(), target_type.clone());
-                ctx.tell(
-                    &target,
-                    SpecMessage::with_params(target_action.clone(), state.fields.clone()),
-                )
-                .await;
-            }
-            // Unreachable: construction rejects these via validate_effect_support.
-            // Failing loudly here keeps a table that bypassed construction from
-            // silently dropping effects.
-            temper_jit::table::Effect::ScheduleAction { .. }
-            | temper_jit::table::Effect::ScheduleAtAction { .. }
-            | temper_jit::table::Effect::SpawnEntity { .. } => {
-                return Err(ActorError::HandlerFailed(format!(
-                    "effect {effect:?} is not executable by the Postgres actor runtime"
-                )));
-            }
-        }
-        Ok(())
-    }
-}
-
-/// Numeric delta for the by-param counter effects. Mirrors
-/// `temper-server::entity_actor::effects::counter_delta_from_params`:
-/// accepts a non-negative number or numeric string, defaults to 0.
-fn counter_delta_from_params(params: &serde_json::Value, param: &str) -> usize {
-    params
-        .get(param)
-        .and_then(|value| match value {
-            serde_json::Value::Number(number) => number.as_u64().map(|v| v as usize),
-            serde_json::Value::String(text) => text.parse::<usize>().ok(),
-            _ => None,
-        })
-        .unwrap_or(0)
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    const SIMPLE_SPEC: &str = r#"
-[automaton]
-name = "TestActor"
-states = ["Idle", "Running"]
-initial = "Idle"
-
-[[state]]
-name = "rounds"
-type = "counter"
-initial = "0"
-
-[[action]]
-name = "Start"
-kind = "input"
-from = ["Idle"]
-to = "Running"
-effect = [{ type = "increment", var = "rounds" }]
-
-[[action]]
-name = "Stop"
-kind = "input"
-from = ["Running"]
-to = "Idle"
-"#;
-
-    #[test]
-    fn test_spec_driven_actor_initial_state() {
-        let actor = SpecDrivenActor::from_ioa(SIMPLE_SPEC, HashMap::new()).unwrap();
-        let state_bytes = actor.initial_state();
-        let state: SpecActorState = serde_json::from_slice(&state_bytes).unwrap();
-        assert_eq!(state.status, "Idle");
-        assert_eq!(state.counters.get("rounds"), Some(&0usize));
-    }
-
-    /// Spec exercising every param-driven state effect plus a guard that
-    /// reads the list those effects maintain.
-    const PARAM_EFFECTS_SPEC: &str = r#"
-[automaton]
-name = "Inventory"
-states = ["Active", "Locked"]
-initial = "Active"
-
-[[state]]
-name = "tags"
-type = "list"
-
-[[state]]
-name = "total"
-type = "counter"
-initial = "0"
-
-[[state]]
-name = "progress"
-type = "counter"
-initial = "0"
-
-[[action]]
-name = "AddTag"
-kind = "input"
-from = ["Active"]
-to = "Active"
-effect = [{ type = "list_append", var = "tags" }]
-
-[[action]]
-name = "RemoveTag"
-kind = "input"
-from = ["Active"]
-to = "Active"
-effect = [{ type = "list_remove_at", var = "tags" }]
-
-[[action]]
-name = "RecordBatch"
-kind = "input"
-from = ["Active"]
-to = "Active"
-effect = [{ type = "increment", var = "total", amount = "qty" }]
-
-[[action]]
-name = "ConsumeBatch"
-kind = "input"
-from = ["Active"]
-to = "Active"
-effect = [{ type = "decrement", var = "total", amount = "qty" }]
-
-[[action]]
-name = "SetProgress"
-kind = "input"
-from = ["Active"]
-to = "Active"
-effect = [{ type = "set_counter_from_param", var = "progress", param = "value" }]
-
-[[action]]
-name = "Lock"
-kind = "input"
-from = ["Active"]
-to = "Locked"
-guard = "list_length_min tags 1"
-"#;
-
-    /// Send one action to the actor and return the updated state.
-    async fn drive(
-        actor: &SpecDrivenActor,
-        state: &mut Vec<u8>,
-        action: &str,
-        params: serde_json::Value,
-    ) -> SpecActorState {
-        use prost::Message as _;
-
-        let handle = ActorHandle::new("test-ns", actor.actor_type());
-        let ctx = ActorContext::new(handle.clone(), None, None);
-        let spec_msg = SpecMessage::with_params(action, params);
-        let message = Message {
-            id: 1,
-            from: None,
-            to: handle,
-            message_type: "SpecMessage".to_string(),
-            payload: spec_msg.encode_to_vec(),
-            correlation_id: None,
-            created_at: chrono::Utc::now(),
-        };
-        actor
-            .handle(&ctx, state, &message)
-            .await
-            .expect("handle should succeed");
-        serde_json::from_slice(state).expect("state should deserialize")
-    }
-
-    #[tokio::test]
-    async fn param_driven_effects_mutate_state() {
-        let actor = SpecDrivenActor::from_ioa(PARAM_EFFECTS_SPEC, HashMap::new())
-            .expect("spec should build");
-        let mut state = actor.initial_state();
-
-        let s = drive(
-            &actor,
-            &mut state,
-            "AddTag",
-            serde_json::json!({"tags": "urgent"}),
-        )
-        .await;
-        assert_eq!(
-            s.lists.get("tags").map(Vec::as_slice),
-            Some(["urgent".to_string()].as_slice()),
-            "list_append must append the param value"
-        );
-
-        let s = drive(
-            &actor,
-            &mut state,
-            "RecordBatch",
-            serde_json::json!({"qty": 5}),
-        )
-        .await;
-        assert_eq!(
-            s.counters.get("total"),
-            Some(&5),
-            "increment-by-param must add the param delta"
-        );
-
-        let s = drive(
-            &actor,
-            &mut state,
-            "ConsumeBatch",
-            serde_json::json!({"qty": 2}),
-        )
-        .await;
-        assert_eq!(
-            s.counters.get("total"),
-            Some(&3),
-            "decrement-by-param must subtract the param delta"
-        );
-
-        let s = drive(
-            &actor,
-            &mut state,
-            "SetProgress",
-            serde_json::json!({"value": 42}),
-        )
-        .await;
-        assert_eq!(
-            s.counters.get("progress"),
-            Some(&42),
-            "set_counter_from_param must set the counter"
-        );
-
-        let s = drive(
-            &actor,
-            &mut state,
-            "RemoveTag",
-            serde_json::json!({"tags_index": 0}),
-        )
-        .await;
-        assert_eq!(
-            s.lists.get("tags").map(Vec::len),
-            Some(0),
-            "list_remove_at must remove the indexed element"
-        );
-    }
-
-    /// The mis-gating scenario from ARN-179: a guard reading a list that a
-    /// prior transition appended to must see the appended value.
-    #[tokio::test]
-    async fn list_guard_sees_appended_value() {
-        let actor = SpecDrivenActor::from_ioa(PARAM_EFFECTS_SPEC, HashMap::new())
-            .expect("spec should build");
-        let mut state = actor.initial_state();
-
-        drive(
-            &actor,
-            &mut state,
-            "AddTag",
-            serde_json::json!({"tags": "urgent"}),
-        )
-        .await;
-        let s = drive(&actor, &mut state, "Lock", serde_json::json!({})).await;
-        assert_eq!(
-            s.status, "Locked",
-            "list_length_min guard must pass after list_append appended a value"
-        );
-    }
-
-    #[test]
-    fn unexecutable_effects_rejected_at_construction() {
-        for (label, effect) in [
-            (
-                "schedule",
-                r#"{ type = "schedule", action = "Expire", delay_seconds = 60 }"#,
-            ),
-            (
-                "schedule_at",
-                r#"{ type = "schedule_at", action = "Expire", field = "expires_at" }"#,
-            ),
-            (
-                "spawn",
-                r#"{ type = "spawn", entity_type = "Child", entity_id_source = "{uuid}" }"#,
-            ),
-        ] {
-            let spec = format!(
-                r#"
-[automaton]
-name = "Scheduler"
-states = ["Active"]
-initial = "Active"
-
-[[action]]
-name = "Arm"
-kind = "input"
-from = ["Active"]
-to = "Active"
-effect = [{effect}]
-"#
-            );
-            let result = SpecDrivenActor::from_ioa(&spec, HashMap::new());
-            let error = match result {
-                Ok(_) => {
-                    panic!("{label} effect must be rejected at construction, not silently dropped")
-                }
-                Err(error) => error,
-            };
-            assert!(
-                error.contains(label),
-                "rejection message must name the unsupported {label} effect, got: {error}"
-            );
-        }
-    }
-
-    #[test]
-    fn unrouted_trigger_rejected_at_construction() {
-        const TRIGGER_SPEC: &str = r#"
-[automaton]
-name = "Notifier"
-states = ["Active"]
-initial = "Active"
-
-[[action]]
-name = "Notify"
-kind = "input"
-from = ["Active"]
-to = "Active"
-effect = [{ type = "trigger", name = "SendNotification" }]
-"#;
-
-        let unrouted = SpecDrivenActor::from_ioa(TRIGGER_SPEC, HashMap::new());
-        assert!(
-            unrouted.is_err(),
-            "a trigger effect with no reaction routing must be rejected at construction"
-        );
-
-        let mut routing = HashMap::new();
-        routing.insert(
-            "SendNotification".to_string(),
-            ("NotifierIntegration".to_string(), "send".to_string()),
-        );
-        let routed = SpecDrivenActor::from_ioa(TRIGGER_SPEC, routing);
-        assert!(
-            routed.is_ok(),
-            "a routed trigger effect must be accepted: {:?}",
-            routed.err()
-        );
-    }
-
-    #[test]
-    fn test_routing_map_builder() {
-        let rules = vec![ReactionRule {
-            name: "a".into(),
-            when: temper_runtime::reaction::ReactionTrigger {
-                entity_type: "Agent".into(),
-                action: Some("PrepareContext".into()),
-                to_state: None,
-            },
-            then: temper_runtime::reaction::ReactionTarget {
-                entity_type: "ContextManager".into(),
-                action: "PrepareContext".into(),
-            },
-            resolve_target: temper_runtime::reaction::TargetResolver::SameId,
-        }];
-
-        let maps = build_routing_maps(&rules);
-        assert_eq!(maps["Agent"]["PrepareContext"].0, "ContextManager");
-        assert_eq!(maps["Agent"]["PrepareContext"].1, "PrepareContext");
-    }
-}
